Regenerate sources with VMMaker.oscogLLP64-nice.1920
Version 1919 incorrectly use PRIxPTR instead of PRIxSQPTR.
The first macro is available in C99 but not included in interp.h.

Notice the unwanted sqInt <-> usqInt changes
Code generation is not fully reproducible.
This is due to type inferencer/inliner which don't allways follow the same path
(probably due to unordered method dictionaries)